Refactor freedom-ldscript-generator
The central idea of this refactor is to capture the "strategy"
abstraction. Each strategy knows which set of memories must
be present in the devicetree for it to be valid, and then the ability to
decide which memory to use for each purpose (ram, rom, itim, etc) when
generating a linker script.

To extend the functionality of the linker script to a new layout, create
a new strategy.

The ChosenStrategy strategy is also created which allows the `chosen`
DeviceTree node to specify which node to use for each purpose using the
 - metal,ram
 - metal,rom
 - metal,itim
properties.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>